refactor update_product_type_azure_devops

+++ dojo/pipeline.py
@@ -200,10 +200,7 @@ def update_product_type_azure_devops(backend, uid, user=None, social=None, *args
             # Assign global role
             if office_location in settings.AZURE_DEVOPS_OFFICES_LOCATION.split(",")[1]:
                 Global_Role.objects.get_or_create(user=user, role=Role.objects.get(id=Roles.Maintainer))
-            elif (
-                office_location in settings.AZURE_DEVOPS_OFFICES_LOCATION.split(",")[2]
-                or job_title in settings.AZURE_DEVOPS_JOBS_TITLE.split(",")[1]
-            ):
+            elif office_location in settings.AZURE_DEVOPS_OFFICES_LOCATION.split(",")[2]:
                 Global_Role.objects.get_or_create(user=user, role=Role.objects.get(id=Roles.Reader))
 
             # Assign specific role
@@ -275,16 +272,17 @@ def update_product_type_azure_devops(backend, uid, user=None, social=None, *args
                                     "Deleting membership of user %s from product type %s", user, product_type_name
                                 )
                     else:
-                        clean_project_type_user(user_product_types_names, user)
+                        clean_project_type_user(user_product_types_names, user, job_title)
                 else:
-                    clean_project_type_user(user_product_types_names, user)
+                    clean_project_type_user(user_product_types_names, user, job_title)
 
 
-def clean_project_type_user(user_product_types_names, user):
-    for product_type_name in user_product_types_names:
-        product_type = Product_Type.objects.get(name=product_type_name)
-        Product_Type_Member.objects.filter(product_type=product_type, user=user).delete()
-        logger.debug("Deleting membership of user %s from product type %s", user, product_type_name)
+def clean_project_type_user(user_product_types_names, user, job_title):
+    if job_title not in settings.AZURE_DEVOPS_JOBS_TITLE.split(",")[1]:
+        for product_type_name in user_product_types_names:
+            product_type = Product_Type.objects.get(name=product_type_name)
+            Product_Type_Member.objects.filter(product_type=product_type, user=user).delete()
+            logger.debug("Deleting membership of user %s from product type %s", user, product_type_name)
 
 
 def custom_filter_group(result, graph_client, regex):
